requirements.txt
keyboard==0.13.5
pyperclip==1.9.0
pyautogui==0.9.54
requests==2.32.3
pyinstaller==6.10.0
janome==0.5.0
pystray==0.19.5
Pillow==11.0.0